fix(deps): Update dependencies and fix tests
Downgraded Android Gradle Plugin from 8.6.1 to 8.5.0 in zeapp. This was done to resolve warnings related to Kotlin Multiplatform.

Updated the requests library from version 2.32.3 to 2.32.4 in zehardware/zefirmware. This is a patch version update and likely includes bug fixes or minor improvements.

Refactored OkHttp MediaType.parse() calls to the toMediaType() extension function in Dalle.kt and Gemini.kt for improved conciseness.